manager/requirements.txt
fastapi
pyjwt
cryptography
